refactor : AI 서버로부터 데이터를 수신하는 로직 수정 Spring Webflux -> Spring MVC
