+++ src/main/java/com/example/accidentdetectionservice/domain/accident/dto/AccidentRequestDto.java
@@ -1,6 +1,5 @@
-package com.example.accidentdetectionservice.domain.webflux.dto;
+package com.example.accidentdetectionservice.domain.accident.dto;
 
-import java.util.Arrays;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -8,9 +7,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AccidentRequestDto {
 
-    // webflux 에서 데이터를 주고받을 대는 주로 json 형식 사용
     private Boolean accident;
-    private String pngData; // png 파일의 데이터를 저장할 필드
     private String latitude; // 위도
     private String longitude; // 경도
     private String date;
