+++ src/main/java/com/example/accidentdetectionservice/domain/accident/repository/AccidentRepository.java
@@ -1,6 +1,6 @@
-package com.example.accidentdetectionservice.domain.hospital.repository;
+package com.example.accidentdetectionservice.domain.accident.repository;
 
-import com.example.accidentdetectionservice.domain.hospital.entity.Accident;
+import com.example.accidentdetectionservice.domain.accident.entity.Accident;
 import com.example.accidentdetectionservice.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
