+++ src/main/java/com/example/accidentdetectionservice/domain/hospital/repository/HospitalRepository.java
@@ -1,6 +1,6 @@
 package com.example.accidentdetectionservice.domain.hospital.repository;
 
-import com.example.accidentdetectionservice.domain.hospital.entity.Accident;
+import com.example.accidentdetectionservice.domain.accident.entity.Accident;
 import com.example.accidentdetectionservice.domain.hospital.entity.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
